# artifacts/api-server/src/routes/stores.ts
import { Router, type IRouter, type Request } from "express";
import { eq, and, isNull } from "drizzle-orm";
import { db, storesTable, sessionsTable, conversationsTable, shopKnowledgeTable, withAdminBypass } from "@workspace/db";
import type { Store } from "@workspace/db/schema";
import {
  CreateStoreBody,
  UpdateStoreBody,
  GetStoreParams,
  GetStoreResponse,
  UpdateStoreParams,
  UpdateStoreResponse,
  ListStoresResponse,
  DeleteStoreParams,
} from "@workspace/api-zod";
import { validateMerchantAuth, validateMerchantAuthForStoreList, clearMerchantSessionCache } from "../middleware";
import { invalidateStoreCache, getCachedStorePublicInfo } from "../middleware";
import { encrypt } from "../services/encryption";
import { invalidateToolsListCache } from "../services/mcp-client";
import { invalidateSessionCacheForDomain } from "../middleware";
import { invalidateKnowledgeCache } from "../services/knowledge-cache";
import { sendError, sendZodError } from "../lib/error-response";
import { logAuditFromRequest } from "../services/audit-logger";

const router: IRouter = Router();

type ProviderValue = "openai" | "anthropic" | "xai";

interface StoreResponse {
  storeDomain: string;
  storefrontToken: string | null;
  provider: string;
  model: string;
  hasApiKey: boolean;
  ucpCompliant: boolean;
  chatEnabled: boolean;
  embedEnabled: boolean;
  guardSensitivity: string;
  blockedTopics: string[];
  brandVoice: Store["brandVoice"] | null;
  customInstructions: string | null;
  welcomeMessage: string | null;
  recommendationStrategy: string;
  dataRetentionDays: number;
  createdAt: Date;
}

function storeToResponse(store: Store): StoreResponse {
  return {
    storeDomain: store.storeDomain,
    storefrontToken: store.storefrontToken,
    provider: store.provider,
    model: store.model,
    hasApiKey: !!store.apiKey,
    ucpCompliant: store.ucpCompliant,
    chatEnabled: store.chatEnabled,
    embedEnabled: store.embedEnabled,
    guardSensitivity: store.guardSensitivity,
    blockedTopics: store.blockedTopics,
    brandVoice: store.brandVoice ?? null,
    customInstructions: store.customInstructions ?? null,
    welcomeMessage: store.welcomeMessage ?? null,
    recommendationStrategy: store.recommendationStrategy,
    dataRetentionDays: store.dataRetentionDays,
    createdAt: store.createdAt,
  };
}

router.get("/stores", validateMerchantAuthForStoreList, async (req, res): Promise<void> => {
  const merchantDomain = (req as Request & { merchantStoreDomain: string }).merchantStoreDomain;
  const stores = await db.select().from(storesTable)
    .where(and(eq(storesTable.storeDomain, merchantDomain), isNull(storesTable.deletedAt)))
    .orderBy(storesTable.createdAt);
  res.json(ListStoresResponse.parse(stores.map(storeToResponse)));
});

router.post("/stores", validateMerchantAuthForStoreList, async (req, res): Promise<void> => {
  const parsed = CreateStoreBody.safeParse(req.body);
  if (!parsed.success) {
    sendZodError(res, parsed.error, "POST /stores", req.body);
    return;
  }

  const merchantDomain = (req as Request & { merchantStoreDomain: string }).merchantStoreDomain;
  if (parsed.data.storeDomain !== merchantDomain) {
    sendError(res, 403, "Cannot create store for a different domain");
    return;
  }

  const existing = await db
    .select()
    .from(storesTable)
    .where(and(eq(storesTable.storeDomain, parsed.data.storeDomain), isNull(storesTable.deletedAt)));

  if (existing.length > 0) {
    sendError(res, 409, "Store already exists");
    return;
  }

  let encryptedApiKey: string | undefined;
  if (parsed.data.apiKey) {
    try {
      encryptedApiKey = encrypt(parsed.data.apiKey);
    } catch (err) {
      sendError(res, 503, "Server encryption is not configured. Cannot store API keys.");
      return;
    }
  }

  const [store] = await db
    .insert(storesTable)
    .values({
      storeDomain: parsed.data.storeDomain,
      storefrontToken: parsed.data.storefrontToken,
      provider: parsed.data.provider as ProviderValue,
      model: parsed.data.model,
      apiKey: encryptedApiKey,
    })
    .returning();

  logAuditFromRequest(req, {
    storeDomain: store.storeDomain,
    actor: "merchant",
    action: "store_created",
    resourceType: "store",
    resourceId: store.storeDomain,
  });

  res.status(201).json(GetStoreResponse.parse(storeToResponse(store)));
});

router.get("/stores/:storeDomain", validateMerchantAuth, async (req, res): Promise<void> => {
  const params = GetStoreParams.safeParse(req.params);
  if (!params.success) {
    sendZodError(res, params.error, "GET /stores/:storeDomain", req.params);
    return;
  }

  const [store] = await db
    .select()
    .from(storesTable)
    .where(and(eq(storesTable.storeDomain, params.data.storeDomain), isNull(storesTable.deletedAt)));

  if (!store) {
    sendError(res, 404, "Store not found");
    return;
  }

  res.json(GetStoreResponse.parse(storeToResponse(store)));
});

router.patch("/stores/:storeDomain", validateMerchantAuth, async (req, res): Promise<void> => {
  const params = UpdateStoreParams.safeParse(req.params);
  if (!params.success) {
    sendZodError(res, params.error, "PATCH /stores/:storeDomain", req.params);
    return;
  }

  const parsed = UpdateStoreBody.safeParse(req.body);
  if (!parsed.success) {
    sendZodError(res, parsed.error, "PATCH /stores/:storeDomain body", req.body);
    return;
  }

  const updateData: Partial<Pick<Store, "storefrontToken" | "provider" | "model" | "apiKey" | "ucpCompliant" | "chatEnabled" | "embedEnabled" | "guardSensitivity" | "blockedTopics" | "brandVoice" | "customInstructions" | "welcomeMessage" | "recommendationStrategy" | "dataRetentionDays">> = {};
  if (parsed.data.storefrontToken !== undefined) updateData.storefrontToken = parsed.data.storefrontToken;
  if (parsed.data.provider !== undefined) updateData.provider = parsed.data.provider as ProviderValue;
  if (parsed.data.model !== undefined) updateData.model = parsed.data.model;
  if (parsed.data.apiKey !== undefined) {
    if (parsed.data.apiKey) {
      try {
        updateData.apiKey = encrypt(parsed.data.apiKey);
      } catch (err) {
        sendError(res, 503, "Server encryption is not configured. Cannot store API keys.");
        return;
      }
    } else {
      updateData.apiKey = parsed.data.apiKey;
    }
  }
  if (parsed.data.ucpCompliant !== undefined) updateData.ucpCompliant = parsed.data.ucpCompliant;
  if (parsed.data.chatEnabled !== undefined) updateData.chatEnabled = parsed.data.chatEnabled;
  if (parsed.data.embedEnabled !== undefined) updateData.embedEnabled = parsed.data.embedEnabled;
  if (parsed.data.guardSensitivity !== undefined) updateData.guardSensitivity = parsed.data.guardSensitivity as Store["guardSensitivity"];
  if (parsed.data.blockedTopics !== undefined) updateData.blockedTopics = parsed.data.blockedTopics;
  if (parsed.data.brandVoice !== undefined) updateData.brandVoice = parsed.data.brandVoice as Store["brandVoice"];
  if (parsed.data.customInstructions !== undefined) updateData.customInstructions = parsed.data.customInstructions;
  if (parsed.data.welcomeMessage !== undefined) updateData.welcomeMessage = parsed.data.welcomeMessage;
  if (parsed.data.recommendationStrategy !== undefined) updateData.recommendationStrategy = parsed.data.recommendationStrategy as Store["recommendationStrategy"];
  if ((parsed.data as Record<string, unknown>).dataRetentionDays !== undefined) {
    const val = Number((parsed.data as Record<string, unknown>).dataRetentionDays);
    if (Number.isFinite(val) && val >= 1 && val <= 365) {
      updateData.dataRetentionDays = val;
    }
  }
  const [store] = await db
    .update(storesTable)
    .set(updateData)
    .where(eq(storesTable.storeDomain, params.data.storeDomain))
    .returning();

  if (!store) {
    sendError(res, 404, "Store not found");
    return;
  }

  invalidateStoreCache(params.data.storeDomain);
  if (parsed.data.storefrontToken !== undefined || parsed.data.ucpCompliant !== undefined) {
    invalidateToolsListCache(params.data.storeDomain);
  }

  const changedFields = Object.keys(parsed.data).filter(k => parsed.data[k as keyof typeof parsed.data] !== undefined);
  logAuditFromRequest(req, {
    storeDomain: params.data.storeDomain,
    actor: "merchant",
    action: "store_updated",
    resourceType: "store",
    resourceId: params.data.storeDomain,
    metadata: {
      changedFields,
      apiKeyChanged: parsed.data.apiKey !== undefined,
    },
  });

  res.json(UpdateStoreResponse.parse(storeToResponse(store)));
});

router.get("/stores/:storeDomain/public", async (req, res): Promise<void> => {
  const { storeDomain } = req.params;
  if (!storeDomain) {
    sendError(res, 400, "Missing storeDomain");
    return;
  }

  const store = await getCachedStorePublicInfo(storeDomain);

  if (!store) {
    sendError(res, 404, "Store not found");
    return;
  }

  const VISION_CAPABLE_PROVIDERS = new Set(["openai", "gemini"]);
  const visionSupported = VISION_CAPABLE_PROVIDERS.has(store.provider);

  res.json({
    storeDomain: store.storeDomain,
    chatEnabled: store.chatEnabled,
    welcomeMessage: store.welcomeMessage ?? null,
    visionSupported,
  });
});

router.delete("/stores/:storeDomain", validateMerchantAuth, async (req, res): Promise<void> => {
  const params = DeleteStoreParams.safeParse(req.params);
  if (!params.success) {
    sendZodError(res, params.error, "DELETE /stores/:storeDomain", req.params);
    return;
  }

  const now = new Date();
  await withAdminBypass(async (scopedDb) => {
    await scopedDb.delete(sessionsTable).where(eq(sessionsTable.storeDomain, params.data.storeDomain));
  });
  await db.update(storesTable).set({ deletedAt: now }).where(eq(storesTable.storeDomain, params.data.storeDomain));
  await db.update(conversationsTable).set({ deletedAt: now }).where(and(eq(conversationsTable.storeDomain, params.data.storeDomain), isNull(conversationsTable.deletedAt)));
  await db.update(shopKnowledgeTable).set({ deletedAt: now }).where(and(eq(shopKnowledgeTable.storeDomain, params.data.storeDomain), isNull(shopKnowledgeTable.deletedAt)));
  invalidateStoreCache(params.data.storeDomain);
  invalidateSessionCacheForDomain(params.data.storeDomain);
  clearMerchantSessionCache();
  invalidateKnowledgeCache(params.data.storeDomain);
  invalidateToolsListCache(params.data.storeDomain);

  logAuditFromRequest(req, {
    storeDomain: params.data.storeDomain,
    actor: "merchant",
    action: "store_deleted",
    resourceType: "store",
    resourceId: params.data.storeDomain,
  });

  res.sendStatus(204);
});

export default router;