# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Shopify MCP AI Shopping Agent API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /stores:
    get:
      operationId: listStores
      tags: [stores]
      summary: List all configured stores
      responses:
        "200":
          description: List of stores
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Store"
    post:
      operationId: createStore
      tags: [stores]
      summary: Add a new store
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateStoreBody"
      responses:
        "201":
          description: Store created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Store"
  /stores/{storeDomain}:
    get:
      operationId: getStore
      tags: [stores]
      summary: Get store details
      parameters:
        - name: storeDomain
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Store details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Store"
        "404":
          description: Store not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateStore
      tags: [stores]
      summary: Update store settings
      parameters:
        - name: storeDomain
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateStoreBody"
      responses:
        "200":
          description: Store updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Store"
        "404":
          description: Store not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteStore
      tags: [stores]
      summary: Remove a store
      parameters:
        - name: storeDomain
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Store deleted
  /stores/{storeDomain}/public:
    get:
      operationId: getStorePublic
      tags: [stores]
      summary: Get public store info (no auth required)
      parameters:
        - name: storeDomain
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Public store info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StorePublicInfo"
        "404":
          description: Store not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    Store:
      type: object
      properties:
        storeDomain:
          type: string
        storefrontToken:
          type: ["string", "null"]
        provider:
          type: string
          enum: [openai, anthropic, xai, gemini]
        model:
          type: string
        hasApiKey:
          type: boolean
        ucpCompliant:
          type: boolean
        chatEnabled:
          type: boolean
        embedEnabled:
          type: boolean
        guardSensitivity:
          type: string
          enum: [off, low, medium, high]
        blockedTopics:
          type: array
          items:
            type: string
        brandVoice:
          type: ["object", "null"]
          properties:
            tone:
              type: string
              enum: [friendly, professional, playful, luxury]
            personality:
              type: string
            greeting:
              type: string
            signOff:
              type: string
        customInstructions:
          type: ["string", "null"]
        welcomeMessage:
          type: ["string", "null"]
        recommendationStrategy:
          type: string
          enum: [bestsellers_first, new_arrivals_first, price_low_to_high, personalized]
        createdAt:
          type: string
          format: date-time
      required:
        - storeDomain
        - provider
        - model
        - hasApiKey
        - ucpCompliant
        - chatEnabled
        - guardSensitivity
        - blockedTopics
        - recommendationStrategy
        - createdAt
    StorePublicInfo:
      type: object
      properties:
        storeDomain:
          type: string
        chatEnabled:
          type: boolean
        welcomeMessage:
          type: ["string", "null"]
        visionSupported:
          type: boolean
          description: Whether the store's LLM provider supports vision/image input
      required:
        - storeDomain
        - chatEnabled
        - visionSupported
    CreateStoreBody:
      type: object
      properties:
        storeDomain:
          type: string
        storefrontToken:
          type: string
        provider:
          type: string
          enum: [openai, anthropic, xai, gemini]
        model:
          type: string
        apiKey:
          type: string
      required:
        - storeDomain
        - storefrontToken
        - provider
        - model
        - apiKey
    UpdateStoreBody:
      type: object
      properties:
        storefrontToken:
          type: string
        provider:
          type: string
          enum: [openai, anthropic, xai, gemini]
        model:
          type: string
        apiKey:
          type: string
        ucpCompliant:
          type: boolean
        chatEnabled:
          type: boolean
        embedEnabled:
          type: boolean
        guardSensitivity:
          type: string
          enum: [off, low, medium, high]
        blockedTopics:
          type: array
          items:
            type: string
        brandVoice:
          type: ["object", "null"]
          properties:
            tone:
              type: string
              enum: [friendly, professional, playful, luxury]
            personality:
              type: string
              maxLength: 500
            greeting:
              type: string
              maxLength: 200
            signOff:
              type: string
              maxLength: 200
        customInstructions:
          type: ["string", "null"]
          maxLength: 2000
        welcomeMessage:
          type: ["string", "null"]
          maxLength: 500
        recommendationStrategy:
          type: string
          enum: [bestsellers_first, new_arrivals_first, price_low_to_high, personalized]
...